+++ src/nextjs/handlers/auth.ts
@@ -3,12 +3,16 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import handelCallback from './callback';
 import handleUser from './user';
 import handleLogout from './logout';
-import { COOKIE_OPTIONS } from '../utils/constants';
+import { COOKIE_OPTIONS } from '../../shared/utils/constants';
 
-export default function handleAuth(
-  cookieOptions: CookieOptions = COOKIE_OPTIONS
-) {
+export interface HandleAuthOptions {
+  cookieOptions?: CookieOptions;
+  logout?: { returnTo?: string };
+}
+
+export default function handleAuth(options: HandleAuthOptions = {}) {
   return async (req: NextApiRequest, res: NextApiResponse): Promise<void> => {
+    const { cookieOptions = COOKIE_OPTIONS, logout } = options;
     let {
       query: { supabase: route }
     } = req;
@@ -17,11 +21,14 @@ export default function handleAuth(
 
     switch (route) {
       case 'callback':
-        return handelCallback(req, res, cookieOptions);
+        return handelCallback(req, res, { cookieOptions });
       case 'user':
-        return await handleUser(req, res, cookieOptions);
+        return await handleUser(req, res, { cookieOptions });
       case 'logout':
-        return await handleLogout(req, res, cookieOptions);
+        return handleLogout(req, res, {
+          cookieOptions,
+          ...logout
+        });
       default:
         res.status(404).end();
     }